Projeto Finalizado

+++ CamposDealer.Application/InputModels/UpdateSaleInputModel.cs
@@ -9,7 +9,7 @@ namespace CamposDealer.Application.InputModels
         public int IdClient { get; set; }
         public int IdProduct { get; set; }
         public int SalesQtd { get; set; }
-        public int ValueUnitValue { get; set; }
+        public float ValueUnitValue { get; set; }
         public float TotalSaleValue { get; set; }
     }
 }
